+++ src/modules/turmas/services/GetTurmaDetalhadaService.ts
@@ -2,31 +2,25 @@ import { prisma } from "../../../shared/database/prisma";
 import { AppError } from "../../../shared/errors/AppError";
 
 export class GetTurmaDetalhadaService {
-
   async execute(id: number) {
-
-    const turma =
-      await prisma.turma.findUnique({
-        where: {
-          id
+    const turma = await prisma.turma.findUnique({
+      where: {
+        id,
+      },
+      include: {
+        curriculo: true,
+        alunos: {
+          orderBy: {
+            nome: "asc",
+          },
         },
-        include: {
-          alunos: {
-            orderBy: {
-              nome: "asc"
-            }
-          }
-        }
-      });
+      },
+    });
 
     if (!turma) {
-      throw new AppError(
-        "Turma não encontrada."
-      );
+      throw new AppError("Turma não encontrada.");
     }
 
     return turma;
-
   }
-
 }