+++ src/modules/turmas/services/ListTurmasService.ts
@@ -1,15 +1,14 @@
 import { prisma } from "../../../shared/database/prisma";
 
 export class ListTurmasService {
-
   async execute() {
-
     return prisma.turma.findMany({
+      include: {
+        curriculo: true,
+      },
       orderBy: {
-        nome: "asc"
-      }
+        nome: "asc",
+      },
     });
-
   }
-
 }